+++ simplebot_deltaland/consts.py
@@ -1,7 +1,7 @@
 """Constants"""
 from enum import IntEnum
 
-DATABASE_VERSION = 4
+DATABASE_VERSION = 5
 WORLD_ID = 0
 
 MAX_LEVEL = 3
@@ -9,6 +9,7 @@ STARTING_LEVEL = 1
 STARTING_ATTACK = 1
 STARTING_DEFENSE = 1
 STARTING_GOLD = 0
+STARTING_INV_SIZE = 15
 RANKS_REQ_LEVEL = 3
 
 MIN_CAULDRON_GIFT = 20
@@ -51,3 +52,33 @@ class Quality(IntEnum):
     BAD = 0
     NORMAL = 1
     GOOD = 2
+
+
+class EquipmentSlot(IntEnum):
+    BAG = 0
+    HANDS = 1
+    HEAD = 2
+    BODY = 3
+    FEET = 4
+
+
+class Tier(IntEnum):
+    NONE = 0
+    ONE = 1
+    TWO = 2
+    THREE = 3
+    FOUR = 4
+    FIVE = 5
+    SIX = 6
+    SEVEN = 7
+    EIGHT = 8
+    NINE = 9
+    TEN = 10
+
+
+class ItemType(IntEnum):
+    SWORD = 1
+    SHIELD = 2
+
+
+equipable_items = [ItemType.SWORD, ItemType.SHIELD]