+++ CHANGELOG.md
@@ -1,5 +1,14 @@
 # Changelog
 
+## [Unreleased]
+
+- add /inv command to see inventory and equipment
+- add /shop command to visit the shop and allow to /buy and /sell items
+- add /on and /off commands to equip and unequip items
+
 ## v0.1.0
 
 - initial release
+
+
+[Unreleased]: https://github.com/adbenitez/simplebot_deltaland/compare/v0.1.0...HEAD